refactor(console): move steps to individual directories
