+++ gravitee-apim-console-webui/src/management/api/creation/v4/steps/step-1/api-creation-v4-step-1.component.scss
@@ -1,5 +1,5 @@
-@use '@angular/material' as mat;
-@use '@gravitee/ui-particles-angular' as gio;
+@use 'node_modules/@angular/material/index' as mat;
+@use 'node_modules/@gravitee/ui-particles-angular/index' as gio;
 
 .api-creation-v4-step-1 {
   &__row {
