+++ gravitee-apim-console-webui/src/management/api/creation/v4/steps/step-1/api-creation-v4-step-1.harness.ts
