+++ gravitee-apim-console-webui/src/management/api/creation/v4/steps/step-2/api-creation-v4-step-2.harness.ts
